data

+++ src/dance-gen/data/code/filter_split_data.py
@@ -14,37 +14,63 @@ def fileToList(f):
     out = open(f, "r").readlines()
     out = [x.strip() for x in out]
     out = [x for x in out if len(x)]
-    return out
+    return out 
+
+def get_data_lists(split_dir, split_name):
+    filter_list = set(fileToList(Path(split_dir, "ignore_list.txt")))
+    data_list = set(fileToList(Path(split_dir, f"{split_name}_list.txt")))
+    return data_list, filter_list
+
+def get_pair_names_by_sequence(sequence_name, dataset):
+    
+    if dataset == "EDGE":
+        return sequence_name, sequence_name
+    
+    if dataset == "FineDance":
+        return sequence_name, sequence_name
+    
+    if dataset == "Aistpp":
+        # Eg.: 'gBR_sBM_cAll_d04_mBR0_ch01' -> 'mBR0'
+        return sequence_name, sequence_name.split("_")[4]
 
 def split_data_generator(
     dataset_path, 
-    dataset: Literal["EDGE", "FineDance"], 
+    dataset: Literal["EDGE", "FineDance", "Aistpp"], 
     split_name: Literal["train", "test"]
 ):
     # `dataset_path`: diretory containing the raw data folder.
     # train - test split
     # A generator yields tuples of (Sequence name, Motion file path, Wav file path)
+    
     if dataset == "EDGE":
-        raw_motion_path = Path(dataset_path, "edge_aistpp/motions")
-        raw_wav_path = Path(dataset_path, "edge_aistpp/wavs")
+        raw_motion_path = Path(dataset_path, "edge_aistpp", "motions")
+        raw_wav_path = Path(dataset_path, "edge_aistpp", "wavs")
         ext = ".pkl"
-        filter_list = set(fileToList(Path(dataset_path, "edge_aistpp/splits/ignore_list.txt")))
-        data_list = set(fileToList(Path(dataset_path, f"edge_aistpp/splits/{split_name}_list.txt")))
+        split_dir = Path(dataset_path, "edge_aistpp", "splits")
         
     elif dataset == "FineDance":
-        raw_motion_path = Path(dataset_path, "finedance/motion")
-        raw_wav_path = Path(dataset_path, "finedance/music_wav")
+        raw_motion_path = Path(dataset_path, "finedance", "motion")
+        raw_wav_path = Path(dataset_path, "finedance", "music_wav")
         ext = ".npy"
-        filter_list = set(fileToList(Path(dataset_path, "finedance/splits/ignore_list.txt")))
-        data_list = set(fileToList(Path(dataset_path, f"finedance/splits/{split_name}_list.txt")))
+        split_dir = Path(dataset_path, "finedance", "splits")
+    
+    elif dataset == "Aistpp":
+        raw_motion_path = Path(dataset_path, "aistpp", "motions")
+        raw_wav_path = Path(dataset_path, "aistpp", "wav")
+        ext = ".pkl"
+        split_dir = Path(dataset_path, "aistpp", "splits")
+         
     else:
-        raise ValueError(f"No dataset named '{dataset}' found! Choose one among ['EDGE', 'FineDance'].")
+        raise ValueError(f"No dataset named '{dataset}' found! Choose one among ['EDGE', 'FineDance', 'Aistpp'].")
+    data_list, filter_list = get_data_lists(split_dir, split_name)
     
     for sequence_name in sorted(data_list):
         if sequence_name in filter_list:
             continue 
-        motion_file = Path(raw_motion_path, sequence_name + ext)
-        wav_file = Path(raw_wav_path, sequence_name + ".wav")
+        
+        motion_name, wav_name = get_pair_names_by_sequence(sequence_name, dataset)
+        motion_file = Path(raw_motion_path, motion_name + ext)
+        wav_file = Path(raw_wav_path, wav_name + '.wav')
         
         if not (motion_file.is_file() and wav_file.is_file()):
             continue
@@ -79,6 +105,10 @@ def preprocess_FineDance(motion_file):
     return np.concatenate([position, rotation_aa], axis=-1)
 
 
+def preprocess_Aistpp(motion_file):
+    return preprocess_EDGE(motion_file)
+
+
 
 def split_data_files(raw_data_path, dataset_name, split_name):
     """ 
@@ -86,7 +116,7 @@ def split_data_files(raw_data_path, dataset_name, split_name):
     - 动作文件: (length, 3+j*3) Numpy.ndarray, 3(position) + joint * 3(axis-angle)
     - 音频文件: 复制原始 wav 文件
     """
-    data_path = raw_data_path / split_name
+    data_path = Path(raw_data_path, split_name)
     print(f"Creating {dataset_name} {split_name} set in {data_path.absolute()}:")
     motion_path = data_path / "motion"
     wav_path = data_path / "wav"
@@ -96,18 +126,23 @@ def split_data_files(raw_data_path, dataset_name, split_name):
         preprocess_motion = preprocess_EDGE
     elif dataset_name == "FineDance":
         preprocess_motion = preprocess_FineDance
+    elif dataset_name == "Aistpp":
+        preprocess_motion = preprocess_Aistpp
     else:
         raise NotImplementedError
     
     for (seq_name, motion_file, wav_file) in split_data_generator(raw_data_path, dataset_name, split_name):
-        new_motion_file = motion_path / f"{seq_name}.npy"
+        
+        new_motion_file = motion_path / (motion_file.stem + '.npy')
         print(f"Processing: {new_motion_file}")
         motion = preprocess_motion(motion_file)
         np.save(new_motion_file, motion)
         
-        new_wav_file = wav_path / f"{seq_name}.wav"
-        print(f"Copying   : {new_wav_file}")
-        shutil.copy(wav_file, wav_path)
+        new_wav_file = wav_path / wav_file.name
+        if not new_wav_file.exists():
+            
+            print(f"Copying   : {new_wav_file}")
+            shutil.copy(wav_file, wav_path)
 
 
 
@@ -116,36 +151,4 @@ def split_data_files(raw_data_path, dataset_name, split_name):
 
 
 if __name__ == "__main__":
-    for i, (a, b, c) in enumerate(split_data_generator("/ssd/shenty/EDGE", "EDGE", "train")):
-        print("name:", a)
-        print("motion:", b)
-        print("wav:", c)
-        print()
-        if i == 2: 
-            break
-    
-    
-    # motion_path = Path(dataset_path, split_name, "motions")
-    # wav_path = Path(dataset_path, split_name, "wavs")
-    # motion_path.mkdir(parents=True, exist_ok=True)
-    # wav_path.mkdir(parents=True, exist_ok=True)
-    
-    
-    # for split_list, split_name in zip([train_list, test_list], ["train", "test"]):
-    #     new_motion_path = Path(dataset_path, f"{split_name}/motions").mkdir(parents=True, exist_ok=True)
-    #     new_wav_path = Path(dataset_path, f"{split_name}/wavs").mkdir(parents=True, exist_ok=True)
-        
-    #     for sequence in split_list:
-    #         if sequence in filter_list:
-    #             continue
-    #         motion = f"{dataset_path}/motions/{sequence}.pkl"
-    #         wav = f"{dataset_path}/wavs/{sequence}.wav"
-    #         assert os.path.isfile(motion)
-    #         assert os.path.isfile(wav)
-    #         motion_data = pickle.load(open(motion, "rb"))
-    #         trans = motion_data["smpl_trans"]
-    #         pose = motion_data["smpl_poses"]
-    #         scale = motion_data["smpl_scaling"]
-    #         out_data = {"pos": trans, "q": pose, "scale": scale}
-    #         pickle.dump(out_data, open(f"{split_name}/motions/{sequence}.pkl", "wb"))
-    #         shutil.copyfile(wav, f"{split_name}/wavs/{sequence}.wav")
+    split_data_files("/root/autodl-tmp/Aistpp/", "Aistpp", "test")
